+++ src/assets/imgs/q7.png
