+++ src/components/Login/Logo.jsx
@@ -1,15 +1,16 @@
 import React from "react";
 import { ColContainer } from "../commons/Container";
 import styled from "styled-components";
-
+import { ReactComponent as ZIcon } from "@assets/imgs/Z.svg";
+import LogoImg from "@assets/imgs/logo.png";
 const Logo = () => {
 	return (
 		<LogoWrapper>
 			<SubLogo>
-				<img src={"./img/z.png"} width="16px" alt="z문자" />
+				<ZIcon width="16px" alt="z문자" />
 				세대 멸망 대비 프로젝트
 			</SubLogo>
-			<img src="./img/logo.png" width="250px" id="logo" alt="logo" />
+			<img src={LogoImg} width="250px" id="logo" alt="logo" />
 		</LogoWrapper>
 	);
 };
