+++ src/pages/Login.jsx
@@ -2,13 +2,14 @@ import React from "react";
 import styled from "styled-components";
 import { ColContainer } from "../components/commons/Container";
 import KakaoBtn from "../components/commons/KakaoBtn";
-import LogoWrapper from "../components/Login/Logo";
+import Logo from "../components/Login/Logo";
+import RandingImg from "@assets/imgs/randingImg.png";
 
 const Login = () => {
 	return (
 		<WhiteBorderContainer spaceAround>
-			<LogoWrapper />
-			<img src="./img/randingImg.png" width="80%" alt="밀레니엄버그img" />
+			<Logo />
+			<img src={RandingImg} width="80%" alt="밀레니엄버그img" />
 			<KakaoBtn redirectUri={"/myPage"} />
 		</WhiteBorderContainer>
 	);
